--- projects/ui-framework/src/lib/overlays/panel-example/panel-example.component.html
@@ -1,25 +1,0 @@
-<b-panel style="margin-right: 20px;">
-
-  <b-button panel-trigger>
-    open Syria
-  </b-button>
-
-  <div panel-content>
-    <h1>Syria</h1>
-    <p>Syria (Arabic: سوريا‎ Sūriyā), officially the Syrian Arab Republic (Arabic: الجمهورية العربية السورية‎ al-Jumhūrīyah al-ʻArabīyah as-Sūrīyah), is a country in Western Asia, bordering Lebanon and the Mediterranean Sea to the west, Turkey to the north, Iraq to the east, Jordan to the south, and Israel to the southwest.</p>
-  </div>
-
-</b-panel>
-
-<b-panel>
-
-  <b-button panel-trigger>
-    open Iraq
-  </b-button>
-
-  <div panel-content>
-    <h1>Iraq</h1>
-    <p>raq (/ɪˈræk/, /ɪˈrɑːk/ (About this soundlisten) or /aɪˈræk/; Arabic: العراق‎ al-'Irāq; Kurdish: عێراق‎ Eraq), officially known as the Republic of Iraq (Arabic: جُمُهورية العِراق‎ ; Kurdish: کۆماری عێراق‎ Komari Eraq), is a country in Western Asia, bordered by Turkey to the north, Iran to the east, Kuwait to the southeast, Saudi Arabia to the south, Jordan to the southwest and Syria to the west.</p>
-  </div>
-
-</b-panel>
