--- projects/ui-framework/src/lib/overlays/panel-example/panel-example.component.ts
@@ -1,16 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'b-panel-example',
-  templateUrl: './panel-example.component.html',
-})
-
-export class PanelExampleComponent implements OnInit {
-
-  constructor(
-  ) {
-  }
-
-  ngOnInit(): void {
-  }
-}
